+++ python_ta/debug/__init__.py
@@ -1,1 +1,2 @@
 from .accumulation_table import AccumulationTable
+from .recursion_table import RecursionTable