+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `AccumulationTable` now stores deep copies of objects rather than shallow copies, thus fixing issues that come up in case of mutation during loop.
 - `AccumulationTable` can now take in any accumulator expressions, for eg. `x * 2`, instead of just variables.
 
+### Bug fixes
+
+- Fix bug in ending location setting for `Attribute` and `DelAttr` nodes when the same attribute
+  was accessed twice on the same line.
+
 ## [2.6.4] - 2024-11-10
 
 ### Bug fixes