+++ CHANGELOG.md
@@ -14,8 +14,9 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added snapshot function for deriving a list of dictionaries containing local variables from relevant functions and/or stack frames.
 - Added new configuration option `allow-pylint-comments` to let users choose whether PythonTA should allow comments beginning with pylint: or not.
 - `AccumulationTable` can now track variables initialized within the `for` loop. Prior, only variables initialized before the `for` loop could be tracked.
-- `AccumulationTable` can now take in any accumulator expressions, for eg. `x * 2`, instead of just variables.
 - `AccumulationTable` now stores deep copies of objects rather than shallow copies, thus fixing issues that come up in case of mutation during loop.
+- `AccumulationTable` can now take in any accumulator expressions, for eg. `x * 2`, instead of just variables.
+- `AccumulationTable` now has an optional initialization argument `output` which allows the users to choose whether they want to write the Accumulation Table to a file.
 - Created a `RecursionTable` context manager for recursive tracing using a tabular output.
 
 ### Bug fixes